# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Personal Finance Advisor API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assumptions/{key}:
    put:
      operationId: updateAssumption
      tags: [assumptions]
      summary: Update a single assumption by key
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAssumptionBody"
      responses:
        "200":
          description: Updated assumption
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assumption"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ─── Balances ──────────────────────────────────────────────────────────────
  /balances:
    get:
      operationId: getBalances
      tags: [balances]
      summary: Get current balances (latest per account)
      responses:
        "200":
          description: Current balances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Balance"
    post:
      operationId: createBalance
      tags: [balances]
      summary: Record a new balance entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBalanceBody"
      responses:
        "201":
          description: Created balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Balance"
  /balances/history:
    get:
      operationId: getBalanceHistory
      tags: [balances]
      summary: Get balance history for all accounts
      parameters:
        - name: account_type
          in: query
          schema:
            type: string
        - name: days
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Balance history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Balance"

  # ─── Bills ─────────────────────────────────────────────────────────────────
  /bills:
    get:
      operationId: getBills
      tags: [bills]
      summary: Get all bills
      responses:
        "200":
          description: All bills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bill"
    post:
      operationId: createBill
      tags: [bills]
      summary: Create a new bill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBillBody"
      responses:
        "201":
          description: Created bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
  /bills/{id}:
    get:
      operationId: getBill
      tags: [bills]
      summary: Get a single bill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateBill
      tags: [bills]
      summary: Update a bill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBillBody"
      responses:
        "200":
          description: Updated bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteBill
      tags: [bills]
      summary: Delete a bill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ─── One-Time Expenses ─────────────────────────────────────────────────────
...
  /commissions:
    get:
      operationId: getCommissions
      tags: [commissions]
      summary: Get commission history
      responses:
        "200":
          description: Commission records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Commission"
    post:
      operationId: createCommission
      tags: [commissions]
      summary: Create a commission entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommissionBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Commission"
  /commissions/{id}:
    patch:
      operationId: updateCommission
      tags: [commissions]
      summary: Update a commission entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCommissionBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Commission"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteCommission
      tags: [commissions]
      summary: Delete a commission entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── Assumptions ─────────────────────────────────────────────────────────
    Assumption:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
        updatedAt:
          type: string
          format: date-time
      required:
        - key
        - value
        - updatedAt
    UpdateAssumptionBody:
      type: object
      properties:
        value:
          type: string
      required:
        - value

    # ─── Balances ─────────────────────────────────────────────────────────────
    Balance:
      type: object
      properties:
        id:
          type: integer
        accountType:
          type: string
        amount:
          type: number
        asOfDate:
          type: string
          format: date-time
        source:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - id
        - accountType
        - amount
        - asOfDate
        - source
    CreateBalanceBody:
      type: object
      properties:
        accountType:
          type: string
          enum: [checking, hysa, brokerage, 401k, other]
        amount:
          type: number
        asOfDate:
          type: string
          format: date-time
        source:
          type: string
          enum: [manual, plaid, csv]
        notes:
          type: ["string", "null"]
      required:
        - accountType
        - amount
        - asOfDate
        - source

    # ─── Bills ───────────────────────────────────────────────────────────────
    Bill:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        amount:
          type: number
        dueDay:
          type: integer
        frequency:
          type: string
        includeInCycle:
          type: boolean
        category:
          type: string
        autopay:
          type: boolean
        notes:
          type: ["string", "null"]
        activeFrom:
          type: ["string", "null"]
          format: date
        activeUntil:
          type: ["string", "null"]
          format: date
        countsThisCycle:
          type: boolean
        nextDueDate:
          type: ["string", "null"]
          format: date
      required:
        - id
        - name
        - amount
        - dueDay
        - frequency
        - includeInCycle
        - category
        - autopay
        - countsThisCycle
    CreateBillBody:
      type: object
      properties:
        name:
          type: string
        amount:
          type: number
        dueDay:
          type: integer
        frequency:
          type: string
          enum: [monthly, quarterly, annual, one-time]
        includeInCycle:
          type: boolean
        category:
          type: string
          enum: [essential, discretionary, debt, variable]
        autopay:
          type: boolean
        notes:
          type: ["string", "null"]
        activeFrom:
          type: ["string", "null"]
          format: date
        activeUntil:
          type: ["string", "null"]
          format: date
      required:
        - name
        - amount
        - dueDay
        - frequency
        - includeInCycle
        - category
        - autopay
    UpdateBillBody:
      type: object
      properties:
        name:
          type: string
        amount:
          type: number
        dueDay:
          type: integer
        frequency:
          type: string
        includeInCycle:
          type: boolean
        category:
          type: string
        autopay:
          type: boolean
        notes:
          type: ["string", "null"]
        activeFrom:
          type: ["string", "null"]
          format: date
        activeUntil:
          type: ["string", "null"]
          format: date

    # ─── One-Time Expenses ────────────────────────────────────────────────────
...
    Commission:
      type: object
      properties:
        id:
          type: integer
        salesMonth:
          type: string
          format: date
        mrrAchieved:
          type: number
        nrrAchieved:
          type: number
        mrrPayout:
          type: number
        nrrPayout:
          type: number
        grossTotal:
          type: number
        takeHome:
          type: number
        payoutDate:
          type: ["string", "null"]
          format: date
        status:
          type: string
          enum: [pending, paid, confirmed]
        notes:
          type: ["string", "null"]
      required:
        - id
        - salesMonth
        - mrrAchieved
        - nrrAchieved
        - mrrPayout
        - nrrPayout
        - grossTotal
        - takeHome
        - status
    CreateCommissionBody:
      type: object
      properties:
        salesMonth:
          type: string
          format: date
        mrrAchieved:
          type: number
        nrrAchieved:
          type: number
        status:
          type: string
          enum: [pending, paid, confirmed]
        notes:
          type: ["string", "null"]
      required:
        - salesMonth
        - mrrAchieved
        - nrrAchieved
        - status
    UpdateCommissionBody:
      type: object
      properties:
        salesMonth:
          type: string
          format: date
        mrrAchieved:
          type: number
        nrrAchieved:
          type: number
        status:
          type: string
          enum: [pending, paid, confirmed]
        notes:
          type: ["string", "null"]
...